Google Maps JavaScript API to combine markers(mulitple locations)
Marker Clustering

This sample shows you how to use marker clusters to display a large number of markers on a map. You can use the MarkerClusterer library in combination with the Google Maps JavaScript API to combine markers of close proximity into clusters, and simplify the display of markers on the map.